Explaining backreferences
This was just an asside in the textbook, but the syntax threw me for a loop, so
I included some exposition about how it works and doesn't work.

Also, note to self, BRE vs ERE syntax is definitely something to be careful of.
the syntax threw me off for a while.